chore: upgrade Storybook to v10 and consolidate configuration imports
