+++ apps/storybook/package.json
@@ -17,14 +17,8 @@
   },
   "devDependencies": {
     "@ously/tsconfig": "workspace:*",
-    "@storybook/addon-essentials": "catalog:",
-    "@storybook/addon-interactions": "catalog:",
-    "@storybook/addon-links": "catalog:",
-    "@storybook/addon-viewport": "catalog:",
-    "@storybook/blocks": "catalog:",
-    "@storybook/experimental-nextjs-vite": "catalog:",
+    "@storybook/nextjs-vite": "catalog:",
     "@storybook/react": "catalog:",
-    "@storybook/test": "catalog:",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
     "autoprefixer": "catalog:",
