+++ packages/ui/package.json
@@ -30,8 +30,7 @@
   },
   "devDependencies": {
     "@ously/tsconfig": "workspace:*",
-    "@storybook/react": "8.6.14",
-    "@storybook/test": "8.6.14",
+    "@storybook/react": "10.3.6",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.2",
     "@testing-library/user-event": "^14.6.1",
