+++ pnpm-workspace.yaml
@@ -15,12 +15,6 @@ catalog:
   class-variance-authority: ^0.7.1
   clsx: ^2.1.1
   lucide-react: ^0.378.0
-  storybook: 8.6.14
-  "@storybook/react": 8.6.14
-  "@storybook/addon-essentials": 8.6.14
-  "@storybook/addon-interactions": 8.6.14
-  "@storybook/addon-links": 8.6.14
-  "@storybook/addon-viewport": 8.6.14
-  "@storybook/blocks": 8.6.14
-  "@storybook/experimental-nextjs-vite": 8.6.14
-  "@storybook/test": 8.6.14
+  storybook: 10.3.6
+  "@storybook/react": 10.3.6
+  "@storybook/nextjs-vite": 10.3.6
